+++ sample/app.rb
@@ -11,6 +11,12 @@ require_relative './util'
 class App < Sinatra::Base
   enable :logging
 
+  helpers do
+    def fetch_something(a:, b:, c:)
+      # do something
+    end
+  end
+
   get '/api/me' do
     1.times { Util.do_something }
     2.times { Util.do_something }
@@ -22,6 +28,14 @@ class App < Sinatra::Base
     8.times { Util.do_something }
     9.times { Util.do_something }
     10.times { Util.do_something }
+
+    11.times do
+      fetch_something(
+        a: SecureRandom.uuid,
+        b: SecureRandom.uuid,
+        c: SecureRandom.uuid,
+      )
+    end
 
     json(
       name: "riseshia",